refactor: useBucketCount hook 변경으로 인한 테스트 실패 코드 수정
